Rewrite code intel regression tests (#7874)
