+++ web/package.json
@@ -9,7 +9,6 @@
     "test-e2e": "mocha ./src/e2e/e2e.test.ts",
     "test:regression": "mocha './src/regression/**/*.test.ts'",
     "test:regression:auth": "mocha ./src/regression/auth.test.ts",
-    "test:regression:codenav": "mocha ./src/regression/codenav.test.ts",
     "test:regression:codeintel": "mocha ./src/regression/codeintel.test.ts",
     "test:regression:config-settings": "mocha ./src/regression/config-settings.test.ts",
     "test:regression:core": "mocha ./src/regression/core.test.ts",
