+++ web/src/regression/lsif-data/.gitignore
@@ -1,1 +1,1 @@
-*.lsif
+**/*.lsif
